Refactored image classifier so that common parts can be used by other apps.
