+++ tools/loader/Loader.cpp
@@ -688,9 +688,11 @@ Loader &Loader::registerExtension(std::unique_ptr<LoaderExtension> extension) {
 
 void Loader::postModelLoad(PlaceholderBindings &bindings,
                            ProtobufLoader &protoLoader,
+                           llvm::StringMap<Placeholder *> &placeholderMap,
                            size_t compilationBatchSize) {
   for (auto &&ext : loaderExtensionList_) {
-    ext->postModelLoad(*this, bindings, protoLoader, compilationBatchSize);
+    ext->postModelLoad(*this, bindings, protoLoader, placeholderMap,
+                       compilationBatchSize);
   }
 }
