+++ tools/loader/Loader.h
@@ -57,6 +57,7 @@ class LoaderExtension {
 public:
   /// Called once after ONNX or Caffe2 model loading.
   virtual void postModelLoad(Loader &, PlaceholderBindings &, ProtobufLoader &,
+                             llvm::StringMap<Placeholder *> &,
                              size_t compilationBatchSize) = 0;
   /// Called once at the beginning of the mini-batch inference.
   virtual void inferInitMiniBatch(Loader &, PlaceholderBindings &,
@@ -171,6 +172,7 @@ public:
   Loader &registerExtension(std::unique_ptr<LoaderExtension> ext);
   /// Called once after ONNX or Caffe2 model loading.
   void postModelLoad(PlaceholderBindings &bindings, ProtobufLoader &protoLoader,
+                     llvm::StringMap<Placeholder *> &,
                      size_t compilationBatchSize);
   /// Called at the beginning of each mini-batch inference.
   void inferInitMiniBatch(PlaceholderBindings &bindings, size_t minibatchIndex,
